requirements.txt
pandas>=2.2.2
tqdm>=4.66.4
icecream>=2.1.3
ViennaRNA>=2.6.4
scipy>=1.13.1
statsmodels>=0.14.2
scikit-learn>=1.5.0
imbalanced-learn>=0.12.3
biopython>=1.83
joblib>=1.4.2
